+++ OUDS/Core/Components/Sources/Controls/TextInput/OUDSTextInput.swift
@@ -197,7 +197,7 @@ import SwiftUI
 ///
 /// ![A text input component in light and dark modes with Wireframe theme](component_textInput_Wireframe)
 ///
-/// - Version: 1.3.0 (Figma component design version)
+/// - Version: 1.4.0 (Figma component design version)
 /// - Since: 0.20.0
 @available(iOS 15, macOS 13, visionOS 1, *)
 public struct OUDSTextInput: View {